# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subjects/saved": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get the list of saved subjects",
                "description": "Get the list of saved subjects for the authenticated user.",
                "operationId": "36d55aff0122226eca1a59b5bbf1cae7",
                "parameters": [
                    {
                        "name": "userSaved",
                        "in": "query",
                        "description": "Filter by user saved subjects",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of saved subjects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubjectResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subject/{subject}/save": {
            "post": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Add a course to the user's saved list",
                "description": "Add a course to the authenticated user's saved list.",
                "operationId": "77e38306cc4605599d123f5ae4df6a56",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "The ID of the subject to save",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subject successfully added to saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subject/{subject}/unsave": {
            "delete": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Remove a course from the user's saved list",
                "description": "Remove a course from the authenticated user's saved list.",
                "operationId": "37e5bca94b58d810e1043204534a9433",
                "parameters": [
                    {
                        "name": "subject",
                        "in": "path",
                        "description": "The ID of the subject to unsave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Subject successfully removed from saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Subject not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/user": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update the authenticated user",
                "description": "Update the details of the currently authenticated user.",
                "operationId": "updateUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "User updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpdateUserRequest": {
                "title": "UpdateUserRequest",
                "description": "Request schema for updating user details",
                "properties": {
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1
                    },
                    "email": {
                        "description": "The email of the user",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "minLength": 1
                    },
                    "phone": {
                        "description": "The phone number of the user",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1
                    },
                    "student": {
                        "description": "Student details",
                        "properties": {
                            "name": {
                                "description": "The name of the student",
                                "type": "string",
                                "maxLength": 255,
                                "minLength": 1
                            },
                            "studentNumber": {
                                "description": "The student number",
                                "type": "string",
                                "maxLength": 255,
                                "minLength": 1
                            },
                            "universityId": {
                                "description": "The ID of the university",
                                "type": "integer",
                                "format": "int64"
                            },
                            "gender": {
                                "description": "The gender of the student",
                                "type": "string",
                                "enum": [
                                    "male",
                                    "female",
                                    "other"
                                ]
                            },
                            "birth": {
                                "description": "The birth date of the student",
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "SubjectResource": {
                "title": "Subject Resource",
                "description": "Subject resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the subject",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the subject",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the subject",
                        "type": "string"
                    },
                    "year": {
                        "description": "The academic year of the subject",
                        "type": "integer"
                    },
                    "semester": {
                        "description": "The semester in which the subject is taught",
                        "type": "integer"
                    },
                    "university": {
                        "description": "The university associated with the subject",
                        "type": "object",
                        "nullable": true
                    },
                    "quizezCounts": {
                        "description": "The count of quizzes associated with the subject",
                        "type": "integer",
                        "nullable": true
                    },
                    "questionsCount": {
                        "description": "The count of questions associated with the subject",
                        "type": "integer",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}